# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Sample API Documentation V1",
        "description": " Complete API documentation for Sample application with versioning support",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication V1"
                ],
                "summary": "User Login",
                "description": "Authenticate user and return access token",
                "operationId": "18294b29b61e586ad50e52197084f8e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email or phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "device_type": {
                                        "description": "Type of device used for login",
                                        "type": "string",
                                        "enum": [
                                            "ios",
                                            "android"
                                        ]
                                    },
                                    "device_token": {
                                        "description": "Device token for push notifications",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "user@example.com",
                                "password": "password123",
                                "device_type": "ios",
                                "device_token": "device_token_example"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userinfo": {
                                            "description": "Authenticated user details",
                                            "type": "object"
                                        },
                                        "token": {
                                            "description": "Access token for authenticated user",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Type of the token (e.g., Bearer)",
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "description": "Token expiration time in ISO 8601 format",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message describing the validation issue",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication V1"
                ],
                "summary": "User Logout",
                "description": "Logout user and revoke access token",
                "operationId": "4b88b902a0ef23e6d971c5440c911cbf",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Logout success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}